Fix navigation panel
